pyproject.toml
[project]
name = "data-gathering-api"
version = "0.1.0"
description = "FastAPI service for loading CSVs and querying exam/test data in BigQuery."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.110.2",
    "uvicorn[standard]==0.29.0",
    "google-cloud-bigquery==3.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["functions"]
